update hocky, khoa hoc, 50% tieu chi, lop (ban quan ly))

+++ AppDRL/src/main/java/models/KhoaHoc.java
@@ -52,7 +52,7 @@ public class KhoaHoc {
         this.MamKT = MamKT;
     }
 
-    public void setTGHoc(int TGHoc) {
+    public void setTGHoc(float TGHoc) {
         this.TGHoc = TGHoc;
     }
